add basic article_upload test for APS
